+++ src/Support/NullUserLifecycleHooks.php
@@ -24,6 +24,14 @@ class NullUserLifecycleHooks implements UserLifecycleHooks
     {
     }
 
+    public function beforeUnsuspend(Model $user): void
+    {
+    }
+
+    public function afterUnsuspend(Model $user): void
+    {
+    }
+
     public function beforeDelete(Model $user): void
     {
     }